Do not compile socket activation when USE_SYSTEMD=OFF
Some part was compiled and not used when systemd socket activation
support is disabled. This part uses SO_DOMAIN and SO_PROTOCOL which is
Linux specific (well not available on BSD systems at least).

Socket activation could work without systemd so we might want to
enable it in the future on non-Linux/non-systemd.